+++ test/unit/index.js
@@ -45,6 +45,7 @@ require('./spec/signaling/room');
 
 require('./spec/signaling/v2');
 require('./spec/signaling/v2/renderhintssignaling');
+require('./spec/signaling/v2/publisherhintsignaling');
 require('./spec/signaling/v2/mediasignaling');
 require('./spec/signaling/v2/dominantspeakersignaling');
 require('./spec/signaling/v2/cancelableroomsignalingpromise');